add target without in_app_review for F-Droid
