+++ lib/common/helpers.dart
@@ -15,10 +15,8 @@
 import 'package:cuppa_mobile/common/constants.dart';
 import 'package:cuppa_mobile/common/globals.dart';
 import 'package:cuppa_mobile/data/localization.dart';
-import 'package:cuppa_mobile/data/prefs.dart';
 
 import 'package:flutter/material.dart';
-import 'package:in_app_review/in_app_review.dart';
 import 'package:intl/intl.dart';
 import 'package:region_settings/region_settings.dart';
 
@@ -163,27 +161,4 @@ String formatDenominatorAmount(int i, {required bool useMetric}) {
     isLargeDevice:
         deviceWidth >= largeDeviceSize && deviceHeight >= largeDeviceSize,
   );
-}
-
-// Check if we should ask user to submit an app store review
-void checkReviewPrompt() async {
-  // Only consider prompting if installed from an app store
-  if (packageInfo.installerStore == installSourceAppleStore ||
-      packageInfo.installerStore == installSourceGoogleStore) {
-    // Activity count determines when to prompt
-    int counter = Prefs.reviewPromptCounter;
-    if (counter <= reviewPromptAtCount) {
-      Prefs.incrementReviewPromptCounter();
-
-      // Prompt for review
-      if (counter == reviewPromptAtCount) {
-        final InAppReview inAppReview = InAppReview.instance;
-        if (await inAppReview.isAvailable()) {
-          Future.delayed(promptDelayDuration, () {
-            inAppReview.requestReview();
-          });
-        }
-      }
-    }
-  }
 }
